// src/app/components/mms/model/AdminDashboardGuestInfoModel.ts
export class AdminDashboardGuestInfoDto {
    id: number;
    monthId: number;
    yearId: number;
    employeeId: string;
    tokenCancel: number;
    tokenDate: string;
    tokenGenDate: string;
    tokenNumber: string;
    mealOption: string;
    mealOptionDes: string;
    tokenType: string;
    tokenTypeDes: string;
    monthName: string;
    yearName: string;
    employeeName: string;
    cUnique: string;
    guestId: string;
    hasMealToken: number;
}